wip: update timer and block data
