+++ src/containers/Dashboard/MiningView/components/BlockInfo.tsx
@@ -1,54 +1,9 @@
 import { Stack, Typography, Divider } from '@mui/material';
 
-import { useEffect, useState } from 'react';
-import { useAppStatusStore } from '../../../../store/useAppStatusStore.ts';
+import { useBlockInfo } from '../../../../hooks/useBlockInfo.ts';
 
 function BlockInfo() {
-    const base_node = useAppStatusStore((s) => s.base_node);
-    const blockHeight = base_node?.block_height;
-    const blockTime = base_node?.block_time || 1;
-    const [timeSince, setTimeSince] = useState('');
-
-    useEffect(() => {
-        // Function to calculate the time difference
-        const calculateTimeSince = () => {
-            const now: Date = new Date();
-            const past: Date = new Date(blockTime * 1000); // Convert seconds to milliseconds
-            const diff: number = now.getTime() - past.getTime();
-
-            // Convert the difference to days, hours, minutes, and seconds
-            const days = Math.floor(diff / (1000 * 60 * 60 * 24));
-            const hours = Math.floor(
-                (diff % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60)
-            );
-            const hoursString = hours.toString().padStart(2, '0');
-            const minutes = Math.floor((diff % (1000 * 60 * 60)) / (1000 * 60))
-                .toString()
-                .padStart(2, '0');
-            const seconds = Math.floor((diff % (1000 * 60)) / 1000)
-                .toString()
-                .padStart(2, '0');
-
-            if (days > 0) {
-                setTimeSince(
-                    `${days} day${days === 1 ? '' : 's'}, ${hoursString}:${minutes}:${seconds}`
-                );
-            } else if (hours > 0) {
-                setTimeSince(`${hoursString}:${minutes}:${seconds}`);
-            } else {
-                setTimeSince(`${minutes}:${seconds}`);
-            }
-        };
-
-        // Initial calculation
-        calculateTimeSince();
-
-        // Set interval to update the time since every second
-        const interval = setInterval(calculateTimeSince, 1000);
-
-        // Cleanup interval on component unmount
-        return () => clearInterval(interval);
-    }, [blockTime]);
+    const { timeSince, blockHeight } = useBlockInfo();
 
     return (
         <Stack direction="row" spacing={2}>
@@ -64,9 +19,7 @@ function BlockInfo() {
             <Divider orientation="vertical" flexItem />
             <Stack>
                 <Typography variant="h6">{timeSince}</Typography>
-                <Typography variant="body2">
-                    Current floor build time
-                </Typography>
+                <Typography variant="body2">Current floor build time</Typography>
             </Stack>
         </Stack>
     );
