+++ src/hooks/useSetUp.ts
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useRef } from 'react';
+import { useEffect, useRef } from 'react';
 import { listen } from '@tauri-apps/api/event';
 import { TauriEvent } from '../types.ts';
 import { preload } from '../visuals';
@@ -13,29 +13,24 @@ export function useSetUp() {
     const setSetupDetails = useAppStateStore((s) => s.setSetupDetails);
     const settingUpFinished = useAppStateStore((s) => s.settingUpFinished);
     // TODO: set up separate auto-miner listener
-    const handleSetupListener = useCallback(
-        async () =>
-            listen('message', ({ event: e, payload: p }: TauriEvent) => {
-                console.info('Setup Event:', e, p);
-                switch (p.event_type) {
-                    case 'setup_status':
-                        console.info('Setup status:', p.title, p.progress);
-                        setSetupDetails(p.title, p.progress);
-
-                        if (p.progress >= 1) {
-                            settingUpFinished();
-                            setView('mining');
-                        }
-                        break;
-                    default:
-                        console.warn('Unknown tauri event: ', { e, p });
-                        break;
-                }
-            }),
-        [setSetupDetails, setView, settingUpFinished]
-    );
 
     useEffect(() => {
+        const unlistenPromise = listen('message', ({ event: e, payload: p }: TauriEvent) => {
+            console.info('Setup Event:', e, p);
+            switch (p.event_type) {
+                case 'setup_status':
+                    console.info('Setup status:', p.title, p.progress);
+                    setSetupDetails(p.title, p.progress);
+                    if (p.progress >= 1) {
+                        settingUpFinished();
+                        setView('mining');
+                    }
+                    break;
+                default:
+                    console.warn('Unknown tauri event: ', { e, p });
+                    break;
+            }
+        });
         if (!startupInitiated.current) {
             preload?.();
             startupInitiated.current = true;
@@ -46,7 +41,7 @@ export function useSetUp() {
         }
 
         return () => {
-            handleSetupListener().then((unlisten) => unlisten());
+            unlistenPromise.then((unlisten) => unlisten());
         };
-    }, [handleSetupListener]);
+    }, [setSetupDetails, setView, settingUpFinished]);
 }
